requirements.txt
gradio>=4.0.0
openai>=1.0.0
srt>=3.5.2
python-dotenv>=1.0.0
